refactored kotlin files
