+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="../../../../../../layout/custom_preview.xml" value="0.2655875299760192" />
         <entry key="..\:/Users/Ricardo Sun/Desktop/ECS198F_P3_FoodTruck-/app/src/main/res/layout/fragment_food_truck_reviews.xml" value="0.39947916666666666" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.44565217391304346" />
         <entry key="app/src/main/res/layout/food_item.xml" value="0.44565217391304346" />
