+++ app/src/main/java/com/ecs198f/foodtrucks/adapters/FoodItemListRecyclerViewAdapter.kt
@@ -1,9 +1,10 @@
-package com.ecs198f.foodtrucks
+package com.ecs198f.foodtrucks.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.ecs198f.foodtrucks.databinding.FoodItemBinding
+import com.ecs198f.foodtrucks.models.FoodItem
 
 class FoodItemListRecyclerViewAdapter(private var items: List<FoodItem>):
     RecyclerView.Adapter<FoodItemListRecyclerViewAdapter.ViewHolder>() {
