+++ app/src/main/java/com/ecs198f/foodtrucks/models/FoodReview.kt
@@ -1,4 +1,4 @@
-package com.ecs198f.foodtrucks
+package com.ecs198f.foodtrucks.models
 
 data class FoodReview(
     val id: String,
