+++ app/src/main/res/layout/fragment_food_truck_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FoodTruckDetailFragment">
+    tools:context=".fragments.FoodTruckDetailFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/materialCardView"
