+++ app/src/main/res/layout/fragment_food_truck_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FoodTruckListFragment">
+    tools:context=".fragments.FoodTruckListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/foodTruckListRecyclerView"
